requirements.txt
# Install PyTorch with CUDA 11.8 support first:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
h5py>=3.8.0
pandas>=2.0.0
numpy<2  # Fixed: NumPy 2.x compatibility issue
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
